+++ src/package/themes/default/components/form.ts
@@ -1,0 +1,5 @@
+export default {
+    gap: '0.75rem',
+    textColor: 'var(--vf-text-color)',
+    disabledOpacity: 'var(--vf-states-disabled-opacity)',
+};